Added basic support for an object type.
This works like a vector in other scheme implementations except with an additional "type" value (accessed through the same abstraction-type function as for simpler "abstraction" values).

A simplified (recursive) garbace collection algorithm has been added for when object support is enabled (the previous algorithm was probably better - and can still be used without object support if you want to avoid recursion, but was over-optimised for only dealing with basic pair values).